Move TestClient to a common directory
